Extend btcutil.Address with segregated witness address types
This commit adds two new addresses types: p2wsh and p2wkh with an
encoding as defined within BIP 142.

Although BIP 142 has been marked as “deferred” the new structures
created are of use within the btcd and btcsuite codebase independent of
the newly defined serialization format for the new standard script
templates. Additionally, as addresses are an end-to-end abstraction
which doesn’t exist at the core of Bitcoin, applications may use these
new address types independent of the adoption within the wide
ecosystem.

In the future as new standard address types are drafted and
standardized, the new encodeSegWitAddress can be modified accordingly.